hard-code the SSH private keys for now
This works in both SAM (it did before) and CloufFormation (it did not,
because wherever the private key files ended up in CloudFormation, sshj
could not load them from any of the locations below:
* keyfilename
* src/main/resources/keyfilename (this worked in SAM)
* src\main\resources\keyfilename

With a simple hard-coded key provider the handlers are for the first time
able to authenticate at the development Terraform server from within the
CloudFormation runtime. The create handler times out though, because
Terraform spends too much time launching the example EC2 instance.
Hopefully with a more lightweight test configuration it will be quick
enough to result in a smooth demo of the synchronous SSH back-end and to
move on to other enhancements.

* pom.xml: Remove the <resource> element for the keys.
* BuiltinKeyFile.java: New class to provide the hard-coded private keys.
* TerraformInterfaceSSH.java: Use the new key provider and change the
  localhost username as the public key is in its own test account now.