parameterise synchronous SSH debug controls
Hard-coded TEST_FORCE_SYNCHRONOUS and TEST_FORCE_SYNCHRONOUS_REAL_DELAY
were not good because the user needed to change the code to see how it
runs synchronously, and could have these changes committed and/or
registered in the CloudFormation provider registry later.

Replace these with optional properties of the callback context
(forceSynchronous and disregardCallbackDelay respectively) to make it
straightforward to say "this SAM test makes this synchronous request"
without changing any code. Use that to fix the shipped SAM tests.

Reduce the amount and scope of code in TerraformBaseHandler a bit and
amend logging for clarity.