fixup SAM tests after commit 53f02e5
"Name" is no longer a part of the schema, so remove it for clarity. This
change neither breaks nor fixes the tests, and in my environment the
tests are broken, in that the resource identifier is Java null and the
configuration directory on the TF server is called "null" for every
resource. Defining "Identifier" instead of "Name" seems to hide this
bug, so I can run my tests before it is fixed.