TerraformInterfaceSSH: add support for .zip files
Make uploadConfiguration() try block small again in order not to confuse
the finally block with potential exceptions coming from the helper shell
SSH connections. Add some parameters to it and rename to uploadFile().

Add a replacement uploadConfiguration() method and place the previously
added pathname expansion workaround there as a specific case for plain
text uploads. Add a specific case for ZIP uploads next to it.