bugfix: generate the UUID only once
The create handler needs to generate an identifier for a newly created
resource on the first invocation and to know what that identifier is on
any follow-up invocations for the same resource. The right way of doing
it might be through prevModel, assuming the external caller sets that to
the model returned by the handler before it calls the handler again.
That said, a quick async SAM test run from "cfn invoke" showed that
prevModel was null on the follow-up invocations. Maybe its purpose is to
represent the previous state for the update handler, or the test was not
done right, or "cfn invoke" didn't ping-pong with the handler properly.

Regardless of the value of prevModel, the create handler never used it
in the first place, so each invocation of the create handler generated a
new UUID and used a different directory on the server.

The first response:
{
  "status": "IN_PROGRESS",
  "callbackContext": {
    "stepId": "CREATE_SYNC_MKDIR",
    "lastDelaySeconds": 0,
    "forceSynchronous": false,
    "disregardCallbackDelay": false
  },
  "callbackDelaySeconds": 0,
  "resourceModel": {
    "Identifier": "890b5de0-181b-4733-9bfd-e3939fa9a405",
    "ConfigurationUrl": "http://denis-examples.s3-website.eu-central-1.amazonaws.com/example12.zip"
  }
}

The next response:
    "Identifier": "3ebbadf9-e65b-4591-bb75-0d5b266cff9a",

The next response:
    "Identifier": "8c0ee486-4698-49fa-bd3f-44d5b249c536",

(and so on)

Fix that by placing the current identifier into the callback context.
Maybe this is not the best solution, but it should reliably cover all
possible ways of running this code, which I cannot quickly verify for
prevModel and every edge case such as SAM sync invocation.