eliminate a false positive in ..Nohup::isRunning()
In this context [ -d /proc/`cat %s` ] tests positive both when the
pidfile exists and stands for an existing process (which is the intended
behaviour) and when the "cat" fails, for example, because the pidfile
or the whole working directory does not exist (which is not). Change the
test to eliminate the false positive.

This solves the problem of the resource provider getting stuck in
DELETE_IN_PROGRESS for 4 hours (until CloudFormation times it out) if the
working directory on the server no longer exists (thus isRunning() keeps
mistakenly returning true). Now instead of getting stuck the resource
provider quickly reports a successful DELETE. That said, perhaps it would
be better to report a failure because the Terraform infrastructure might
still exist and the user should be aware of any required manual cleanup.