articulate the error condition in isRunning()
After the previos commit the resource provider would no longer hang
waiting on a non-existent delete process, but the stack would end up in
a DELETE_COMPLETE state without Terraform actually deleting anything,
which was misleading.

Change the shell command to test the cat command before using it in the
command substitution, and to return an error message (meaning it is
impossible to confirm the detached delete process had started and
completed the work) instead of "false" (meaning the detached delete
process is confirmed completed the work) when appropriate.

In particular, this leaves the stack in a DELETE_FAILED state when the
working directory on the Terraform server no longer exists, such that the
user is aware something went wrong and might require a manual cleanup of
the resources.

That said, this solution does not cover 100% of the possible fault space
and can be refined even more in future if required.